move gettime to bin
